mod: remove arena team at top of auto list option
Remove the option for putting the arena team at the top of the
auto generated list. The no longer changeable default is to
ignore the arena team.

This option's importance is as diminished as the squad arena's.
So instead of just making the default to ignore the arena team,
i'm rather reducing the ui clutter and cognitive load by
removing the option.