feat: add ui for the help page
The help page has help sections at the top
(and if appropriate a back button)

Clicking on a help section shows the topics in that section.
Clicking on one of the topics will switch to that topic.
You can get the topics back by clicking the help section again.

You can explore the help page or you can follow a help link,
that will bring you directly to the linked topic.
By hitting the back button located with the help sections,
you can return to the part of the program you were in
before entering the help page.

The welcome message and the full instructions from the optimizer page
are now in the help page.