fix: reset the ui when deleting the last profile
When deleting the last profile with the trashcan button
next to the profile selector, the profile was deleted from the database,
but not completely from redux store. This resulted in the ui still
showing the data of the deleted profile.
Now after deleting the last profile the ui will only show you the parts
that don't need a profile.